Implement support for type-specific views
